Gui editors: Use float for zoom precision (#6008)
This fixes build errors with casts.

+++ include/AutomationEditor.h
@@ -174,7 +174,7 @@ private:
 	ComboBoxModel m_zoomingYModel;
 	ComboBoxModel m_quantizeModel;
 
-	static const QVector<double> m_zoomXLevels;
+	static const QVector<float> m_zoomXLevels;
 
 	FloatModel * m_tensionModel;
